fix: finalize transport retry safeguards
