+++ backend/services/file/upload_concurrency_test.go
@@ -4,11 +4,137 @@ import (
 	"bytes"
 	"context"
 	"errors"
+	"io"
 	"os"
 	"path/filepath"
+	"sync"
 	"testing"
 	"testing/synctest"
+
+	"TDrive/backend/tgclient"
 )
+
+// countingClient tracks concurrent SendFile calls while delegating to Fake.
+type countingClient struct {
+	*tgclient.Fake
+	mu       sync.Mutex
+	inFlight int
+	maxSeen  int
+	calls    int
+	entered  chan<- struct{}
+	release  <-chan struct{}
+}
+
+func (c *countingClient) beginSend() func() {
+	c.mu.Lock()
+	c.inFlight++
+	c.calls++
+	if c.inFlight > c.maxSeen {
+		c.maxSeen = c.inFlight
+	}
+	c.mu.Unlock()
+	return func() {
+		c.mu.Lock()
+		c.inFlight--
+		c.mu.Unlock()
+	}
+}
+
+func (c *countingClient) waitForRelease(ctx context.Context) error {
+	if c.entered != nil {
+		c.entered <- struct{}{}
+	}
+	if c.release == nil {
+		return nil
+	}
+	select {
+	case <-c.release:
+		return nil
+	case <-ctx.Done():
+		return ctx.Err()
+	}
+}
+
+func (c *countingClient) snapshot() (calls, inFlight, maxSeen int) {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	return c.calls, c.inFlight, c.maxSeen
+}
+
+func (c *countingClient) SendFile(ctx context.Context, peer tgclient.InputPeer, r io.Reader, name, caption string, totalSize int64, onProgress func(sent, total int64)) (tgclient.SendFileResult, error) {
+	defer c.beginSend()()
+	if err := c.waitForRelease(ctx); err != nil {
+		return tgclient.SendFileResult{}, err
+	}
+	return c.Fake.SendFile(ctx, peer, r, name, caption, totalSize, onProgress)
+}
+
+func (c *countingClient) SendFileWithRandomID(ctx context.Context, peer tgclient.InputPeer, r io.Reader, name, caption string, totalSize int64, onProgress func(sent, total int64), randomID int64) (tgclient.SendFileResult, error) {
+	defer c.beginSend()()
+	if err := c.waitForRelease(ctx); err != nil {
+		return tgclient.SendFileResult{}, err
+	}
+	return c.Fake.SendFileWithRandomID(ctx, peer, r, name, caption, totalSize, onProgress, randomID)
+}
+
+func TestUploadBatchHonorsConfiguredConcurrency(t *testing.T) {
+	const fileCount = 6
+	for _, tc := range []struct {
+		name    string
+		limit   int
+		wantMax int
+	}{
+		{"serial limit", 1, 1},
+		{"default limit stays bounded", 0, 3},
+	} {
+		t.Run(tc.name, func(t *testing.T) {
+			svc, _, fakeTG, _ := newTestService(t)
+			svc.MaxConcurrentUploads = tc.limit
+
+			var paths, parents []string
+			for i := 0; i < fileCount; i++ {
+				paths = append(paths, writeTempNamedFile(t, "f"+string(rune('a'+i))+".txt", []byte("body")))
+				parents = append(parents, "")
+			}
+			synctest.Test(t, func(t *testing.T) {
+				entered := make(chan struct{}, fileCount)
+				release := make(chan struct{})
+				client := &countingClient{Fake: fakeTG, entered: entered, release: release}
+				svc.TG = client
+
+				resultCh := make(chan []Metadata, 1)
+				errCh := make(chan error, 1)
+				go func() {
+					files, err := svc.Upload(context.Background(), personalChannelID, paths, parents, false)
+					resultCh <- files
+					errCh <- err
+				}()
+
+				for i := 0; i < tc.wantMax; i++ {
+					<-entered
+				}
+				synctest.Wait()
+				calls, inFlight, maxSeen := client.snapshot()
+				if calls != tc.wantMax || inFlight != tc.wantMax || maxSeen != tc.wantMax {
+					t.Fatalf("before release calls/in-flight/max = %d/%d/%d, want %d/%d/%d", calls, inFlight, maxSeen, tc.wantMax, tc.wantMax, tc.wantMax)
+				}
+
+				close(release)
+				synctest.Wait()
+				if err := <-errCh; err != nil {
+					t.Fatalf("upload: %v", err)
+				}
+				if files := <-resultCh; len(files) != fileCount {
+					t.Fatalf("uploaded = %d, want %d", len(files), fileCount)
+				}
+				calls, inFlight, maxSeen = client.snapshot()
+				if calls != fileCount || inFlight != 0 || maxSeen != tc.wantMax {
+					t.Fatalf("final calls/in-flight/max = %d/%d/%d, want %d/0/%d", calls, inFlight, maxSeen, fileCount, tc.wantMax)
+				}
+			})
+		})
+	}
+}
 
 func TestConcurrentMultipartUploadsRecoverIndependentlyAfterSharedFailures(t *testing.T) {
 	svc, _, fakeTG, _ := newTestService(t)
